UIOR-300 load AC units based on permissions
