+++ lib/AcqUnits/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as AcqUnits } from './AcqUnits';
+export { default as AcqUnitsField } from './AcqUnitsField';
