+++ lib/FieldMultiSelection/FieldMultiSelection.js
@@ -8,7 +8,7 @@ import { MultiSelection } from '@folio/stripes/components';
 
 const onBlurDefault = e => { e.preventDefault(); };
 
-const FieldMultiSelection = ({ name, dataOptions, formatter, itemToString, label }) => {
+const FieldMultiSelection = ({ name, dataOptions, formatter, itemToString, label, ...rest }) => {
   return (
     <Field
       component={MultiSelection}
@@ -19,6 +19,7 @@ const FieldMultiSelection = ({ name, dataOptions, formatter, itemToString, label
       formatter={formatter}
       onBlur={onBlurDefault}
       label={label}
+      {...rest}
     />
   );
 };
